Split handler_tests and hint_tests
